fix(Key): updates key styles
